dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Document-createEvent.html

<!DOCTYPE html>
<title>Document.createEvent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testAlias(arg, iface) {
  test(function() {
    if (iface in window) {
      var ev = document.createEvent(arg);
      assert_true(ev instanceof window[iface]);
    }
  }, arg + " should be an alias for " + iface + ".");
}
[
  ["Events", "Event"],
  ["HTMLEvents", "Event"],
  ["MouseEvents", "MouseEvent"],
  ["MutationEvents", "MutationEvent"],
  ["UIEvents", "UIEvent"]
].forEach(function(alias) {
  testAlias(alias[0], alias[1]);
  testAlias(alias[0].toLowerCase(), alias[1]);
  testAlias(alias[0].toUpperCase(), alias[1]);
});

test(function() {
  assert_throws("NOT_SUPPORTED_ERR", function() {
    var evt = document.createEvent("foo");
  });
}, "Should throw NOT_SUPPORTED_ERR for unrecognized arguments");
</script>
